Derek Mead's Wordies quick-start guide for someone who already understands regexes
It might be interesting to have your approach on board, possibly as a
module. Planning the open source project right now - if you end up
converting it to JS, I'll be happy to hear about it. It might even be
used as an UI element to help explain the visual syntax - as it, too, is
terse.

"I've been working on a more human-friendly regex notation for a while.

I have implemented convertors between my 'wordy' notation and
standard(terse) regexes in both directions.

The wordy notation is pure text, so it's editable using whatever you use
for source.
It uses indentation to show structure (like Python), so the structure of
complex regexes is much more obvious.

My current implementation is in Perl: I am planing to port it to
JavaScript. [...]

The document attached is the quick-start guide for someone who already
understands regexes. But the idea is to make regexes friendly enough to
introduce new users -and that would require a book-length document, not
a quick-start.

It might useful as one of several alternative syntaxes: there is no
reason why there can't be automatic converters between all of them
(possible using 'terse' as the common format).

One advantage of a pure-text format is that it can be part of the
source: with JavaScript I could embed it (like XRexExp) so that you
don't have pesky separate conversion utilities and having the wordies
not really part of the code. You can currently store wordies within the
source as comments, but it's untidy.

Feedback welcomed!

Cheers,
Derek"