Wordy web server Windows executable
Hi,

The Wordy<->Terse code is freely distributable. I haven't actually
distributed the source, but its copyright staement is the very
permissive Perl licence.

Cheers,
Derek

Hi Olli,

Attached is a executable (renamed, for mail reasons) that I created
using ActiveState's PerlApp utility.

When you execute it, it starts serving (port 5678 by default, or pass
it a port number on the command line).

The web page it serves is not polished. But the basic functionality is
that you can put a wordy in the Terse textbox andclick 'Terse to
Wordy', or put a wordy in the 'Wordy' textbox and click 'Wordy to
terse'.

The 'Match', 'Replace' and 'Split' buttons take the current wordy and
test it (first converting it to Terse, then executing it).

The execution is done on the server, using the Perl equivalent of the
wordy. But one design aim of wordies is that a given wordy will do the
same thing when used in different languages, so differences between
flavours are mostly coped with. You should be able to choose the Terse
language flavour you want, generate it from a wordy and paste the Terse
regex into some source in your chosen target language.

One note if you are using multi-line test data: the demo website passes
line breaks in the Test Data box as carriage-return / line-feed, which
doesn't match newline: you have to match the CR either explicitly by
carriage-return or generically as whitespace.

The executable is from 2014, but I don't think much has changed since
then :-(

Cheers,
Derek